requirements.txt
Django
djangorestframework
django-rest-knox